# openapi.yaml
openapi: 3.1.0
...
info:
  title: MultiSpeed API
  version: 1.0.3
  summary: Multi-WAN speed-test scheduling, execution, routing, and statistics API
  description: |
    Same-origin API for MultiSpeed. There is deliberately no authentication.
    Keep the listener on loopback or a trusted network unless an authenticating
    reverse proxy protects it. Browser mutations are origin-validated.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: MultiSpeed issue tracker
    url: https://github.com/dude2k/MultiSpeed/issues
servers:
  - url: /api/v1
    description: Same-origin application API
security: []
...
paths:
...
  /providers/ookla/binary:
    get:
      operationId: getOoklaBinaryStatus
      tags: [Providers]
      summary: Report whether managed Ookla executable upload is enabled and installed
      responses:
        "200":
          description: Deployment-owned upload capability and installation state.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OoklaBinaryStatus"
    post:
      operationId: uploadOoklaBinary
      tags: [Providers]
      summary: Atomically validate and install an operator-supplied Ookla executable
      description: |
        Disabled unless APP_ALLOW_OOKLA_BINARY_UPLOAD=true and OOKLA_BINARY resolves
        beneath APP_DATA_DIR. The current Ookla EULA must already be accepted. The
        body must be one Linux amd64 ELF executable obtained separately by the
        operator. MultiSpeed checks its bounded size, architecture, executable
        version output, and only then atomically activates it. This endpoint executes
        the uploaded file for verification and must only be enabled on a trusted
        private network or behind an authenticating reverse proxy.
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              maxLength: 67108864
      responses:
        "201":
          description: The validated executable was installed in managed persistent storage.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OoklaBinaryInstallResult"
        "403":
          description: Managed executable upload is not enabled by deployment policy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          description: The executable exceeds the 64 MiB upload limit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /settings:
    get:
      operationId: getSettings
      tags: [Settings]
      summary: Get persisted application settings
      responses:
        "200":
          description: Current settings.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateSettings
      tags: [Settings]
      summary: Replace persisted application settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Settings"
      responses:
        "200":
          description: Updated settings.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          $ref: "#/components/responses/InvalidJSON"
        "415":
          $ref: "#/components/responses/UnsupportedContentType"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  headers:
    RequestID:
      description: Request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
  responses:
    InvalidJSON:
      description: Malformed, oversized, multiple-value, or schema-incompatible JSON.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: The resource, schedule, target, interface, source address, or route is invalid.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnsupportedContentType:
      description: "A JSON body was sent without Content-Type: application/json."
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RateLimited:
      description: Manual execution or provider discovery rate limit exceeded.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Retry-After:
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Sanitized internal error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code:
              type: string
              pattern: "^[A-Z][A-Z0-9_]*$"
              maxLength: 64
            message:
              type: string
              minLength: 1
              maxLength: 8192
            requestId:
              type: string
              minLength: 1
              maxLength: 128
      examples:
        - error:
            code: ROUTE_VALIDATION_FAILED
            message: The configured route does not use the selected WAN interface.
            requestId: 73c754bea60aa4e18657a537
...
    OoklaBinaryStatus:
      type: object
      additionalProperties: false
      required: [uploadEnabled, installed, maxUploadBytes, message]
      properties:
        uploadEnabled:
          type: boolean
        installed:
          type: boolean
        maxUploadBytes:
          type: integer
          minimum: 1
          maximum: 67108864
        message:
          type: string
          maxLength: 1024
    OoklaBinaryInstallResult:
      type: object
      additionalProperties: false
      required: [uploadEnabled, installed, maxUploadBytes, message, version, sha256]
      properties:
        uploadEnabled:
          type: boolean
        installed:
          const: true
        maxUploadBytes:
          type: integer
          minimum: 1
          maximum: 67108864
        message:
          type: string
          maxLength: 1024
        version:
          type: string
          minLength: 1
          maxLength: 512
        sha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
...
    Settings:
      type: object
      additionalProperties: false
      required:
        - displayUnits
        - defaultTimezone
        - globalConcurrency
        - allowSeparateWanConcurrency
        - retentionMode
        - retentionValue
        - defaultChartRange
        - interfaceRefreshIntervalSeconds
        - defaultTaskTimeoutSeconds
        - databaseMaintenanceSchedule
      properties:
        displayUnits:
          type: string
          enum: [bits, bytes]
        defaultTimezone:
          type: string
          minLength: 1
          maxLength: 128
          examples: [UTC]
        globalConcurrency:
          type: integer
          minimum: 1
          maximum: 16
          default: 1
        allowSeparateWanConcurrency:
          type: boolean
          default: false
        retentionMode:
          type: string
          enum: [forever, days, months]
        retentionValue:
          type: integer
          minimum: 0
          maximum: 3650
        defaultChartRange:
          type: string
          enum: [24h, 7d, 30d, 90d]
        interfaceRefreshIntervalSeconds:
          type: integer
          minimum: 5
          maximum: 3600
        defaultTaskTimeoutSeconds:
          type: integer
          minimum: 5
          maximum: 3600
        databaseMaintenanceSchedule:
          type: string
          minLength: 9
          maxLength: 256
        ooklaEulaAccepted:
          type: boolean
          readOnly: true
          description: Whether an operator recorded explicit acceptance through the dedicated licensing endpoint.
        ooklaEulaAcceptedAt:
          type: [string, "null"]
          format: date-time
          readOnly: true
          description: UTC timestamp of the persisted acceptance, or null while revoked.
        ooklaEulaVersion:
          type: string
          readOnly: true
          description: Terms revision recorded with the persisted acknowledgement, or empty when none is stored.
        ooklaEulaCurrentVersion:
          type: string
          readOnly: true
          description: Terms revision required by this MultiSpeed release.
        ooklaEulaEffectiveAccepted:
          type: boolean
          readOnly: true
          description: Effective technical-gate state after applying the persisted acknowledgement and environment override.
        ooklaEulaAcceptanceSource:
          type: string
          enum: [none, persisted, environment]
          readOnly: true
          description: Source currently opening the technical gate.
...